🎉 add / .code_rabbit.yaml 생성
