--- src/app/(main)/layout.tsx
@@ -1,22 +1,0 @@
-import Header from '@/components/header';
-
-import React, { PropsWithChildren } from 'react';
-import { createClient } from '../utils/supabase/server';
-import { redirect } from 'next/navigation';
-
-const MainLayout = async ({ children }: PropsWithChildren) => {
-  const supabase = await createClient();
-  const { data, error } = await supabase.auth.getUser();
-  // console.log('유저데이터', data.user);
-  if (error || !data?.user) {
-    redirect('/auth/login');
-  }
-  return (
-    <div>
-      <Header />
-      <main className="px-4">{children}</main>
-    </div>
-  );
-};
-
-export default MainLayout;
