--- src/app/(main)/test/page.tsx
@@ -1,5 +1,0 @@
-const TestPage = () => {
-  return <div>TestPage</div>;
-};
-
-export default TestPage;
