+++ src/app/utils/openai/route.ts
@@ -4,7 +4,7 @@ import OpenAI from 'openai';
 import { COMMENT_PROMPT } from '../../../../src/constants/openai/commentConfig';
 const openai = new OpenAI();
 
-export async function POST(req: Request) {
+export const POST = async (req: Request) => {
   const body = await req.json();
   //JSON.stringify(body)로 입력 보내기
   const completion = await openai.chat.completions.create({
@@ -22,4 +22,31 @@ export async function POST(req: Request) {
   });
   // console.log(completion.choices[0].message);
   return NextResponse.json({ content: completion.choices[0].message.content });
-}
+};
+
+//route 사용함수는 사용하는 곳에서 정의한 뒤 사용하기기
+
+export const fetchGPT = async ({
+  topic,
+  emotions,
+  message
+}: {
+  topic: string | null;
+  emotions: string[];
+  message: string;
+}) => {
+  const res = await fetch('/utils/openai', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify({
+      content: `주제: ${topic}, 감정: ${emotions.join(
+        ', '
+      )}, 메시지: ${message}`
+    })
+  });
+
+  const data = await res.json();
+  return data.content;
+};
